project/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.16.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asteroid-filterbanks==0.4.0
attrs==25.3.0
audioread==3.0.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
deep-sort-realtime==1.3.2
docopt==0.6.2
einops==0.8.1
emotiefflib==1.0
fastapi==0.115.13
ffmpeg-python==0.2.0
ffmpy==0.6.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
future==1.0.0
gigaam==0.1.0
gradio==5.34.1
gradio-client==1.10.3
greenlet==3.2.3
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.4
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
humanfriendly==10.0
hydra-core==1.3.2
hyperpyyaml==1.2.2
hypothesis==6.135.11
idna==3.10
jinja2==3.1.6
joblib==1.5.1
julius==0.2.7
kenlm==0.3.0
kiwisolver==1.4.8
lazy-loader==0.4
librosa==0.11.0
lightning==2.5.1.post0
lightning-utilities==0.14.3
llvmlite==0.44.0
mako==1.3.10
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.5.0
# networkx==3.5
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.17.3
opencv-python==4.11.0.86
optuna==4.4.0
orjson==3.10.18
packaging==24.2
pandas==2.3.0
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
primepy==1.3
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyannote-audio==3.3.2
pyannote-core==5.0.0
pyannote-database==5.1.3
pyannote-metrics==3.2.1
pyannote-pipeline==3.0.1
pycparser==2.22
pyctcdecode==0.5.0
pydantic==2.11.7
pydantic-core==2.33.2
pydub==0.25.1
pyee==13.0.0
pygments==2.19.1
pygtrie==2.5.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==2.5.1.post0
pytorch-metric-learning==2.8.1
pytz==2025.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
ruamel-yaml==0.18.14
ruamel-yaml-clib==0.2.12
ruff==0.12.0
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
semantic-version==2.10.0
semver==3.0.4
sentencepiece==0.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==0.5.0.post1
speechbrain==1.0.3
sqlalchemy==2.0.41
starlette==0.46.2
sympy==1.13.1
tabulate==0.9.0
tensorboardx==2.6.4
threadpoolctl==3.6.0
timm==0.9.16
tokenizers==0.21.1
tomlkit==0.13.3
torch==2.5.1
torch-audiomentations==0.12.0
torch-pitch-shift==1.2.5
torchaudio==2.5.1
torchmetrics==1.7.3
torchvision==0.20.1
tqdm==4.67.1
transformers==4.49.0
triton==3.1.0
typer==0.16.0
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
ultralytics==8.3.156
ultralytics-thop==2.0.14
urllib3==2.5.0
uvicorn==0.34.3
websockets==15.0.1
yarl==1.20.1
